feat: 日志记录不再使用 tracing-appender
tracing-appender 会加追日志，更适用于服务端，本地程序日志应在启动程序时清空旧日志，或重命名旧日志，本程序采用清空旧日志的方案。
