+++ src-tauri/Cargo.toml
@@ -24,7 +24,6 @@ tracing-subscriber = { version = "0", features = [
     'env-filter',
     'json',
 ] }
-tracing-appender = '0'
 thiserror = "1"
 sqlx = { version = "0", default-features = false, features = [
     "macros",
@@ -44,13 +43,13 @@ tauri-plugin-shell = "2.0.0-rc"
 tauri-plugin-clipboard-manager = "2.0.0-rc"
 tauri-plugin-dialog = "2.0.0-rc"
 
+[target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
+tauri-plugin-single-instance = "2.0.0-rc"
+tauri-plugin-updater = "2.0.0-rc"
+
 [profile.release]
 panic = "abort"
 codegen-units = 1
 lto = true
 opt-level = "s"
 strip = true
-
-[target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
-tauri-plugin-single-instance = "2.0.0-rc"
-tauri-plugin-updater = "2.0.0-rc"
